修改bug

+++ Service/shop-hapi/src/main/java/com/genyuanlian/consumer/service/IBWSService.java
@@ -2,6 +2,8 @@ package com.genyuanlian.consumer.service;
 
 import com.genyuanlian.consumer.shop.vo.BWSWalletCreateResponseVo;
 
+import java.math.BigDecimal;
+
 /**
  * BWS 接口的二次封装
  * Created by hunter.wang on 2018/5/10.
@@ -20,7 +22,7 @@ public interface IBWSService
      * @param walletId
      * @return
      */
-    public Double walletBalance(String walletId);
+    public BigDecimal walletBalance(String walletId);
 
     /**
      * 钱包充值
@@ -28,7 +30,7 @@ public interface IBWSService
      * @param amount
      * @return
      */
-    public String walletRecharge(Long businessId,Long ownerId,int ownerType,String walletMainAddress,Double amount);
+    public String walletRecharge(Long businessId,Long ownerId,int ownerType,String walletMainAddress,BigDecimal amount);
 
     /**
      * 钱包消费
@@ -36,7 +38,7 @@ public interface IBWSService
      * @param amount
      * @return
      */
-    public String walletConsume(Long businessId,Long ownerId,int ownerType,String walletId,Double amount);
+    public String walletConsume(Long businessId,Long ownerId,int ownerType,String walletId,BigDecimal amount);
